fix: move note meta dtos back to the same file
